feat: Repository 추상화
