+++ domain/src/main/java/woowacourse/shopping/data/ProductRepository.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.data
+
+import woowacourse.shopping.model.Product
+
+interface ProductRepository {
+    fun getAllProducts(): List<Product>
+}
